Added Popup and removed select avail type component
