+++ app/javascript/src/containers/Availability.jsx
@@ -8,7 +8,7 @@ import { bindActionCreators } from 'redux';
 import NavBar from './NavBar';
 
 // components
-import { Card } from 'semantic-ui-react';
+import { Card, Icon, Label, Popup } from 'semantic-ui-react';
 
 import AvailCal from './../components/availability';
 
@@ -21,7 +21,6 @@ import {
   dragDropUpdate,
 } from '../actions/user';
 
-
 class Availability extends Component {
   componentDidMount() {
     this.props.checkSession();
@@ -32,13 +31,42 @@ class Availability extends Component {
       user,
       putAvail, postAvail, deleteAvail, dragDropUpdate,
     } = this.props;
+
+    const popupContent = (
+      <span>
+        When creating your Availability Calendar, you can switch between the options below to help your captain (or whomever is making a shift) know how available you are.
+        <br />
+        <br />
+        <p>
+          <Label circular color="green">Available</Label>:
+          means you are able to tent at this time 100% of the time.
+        </p>
+        <p>
+          <Label circular color="yellow">Somewhat Available</Label>:
+           means you are able to tent, only if necessarry due to other commitments (travelling from East, end of classes, etc.).
+        </p>
+         <p>
+           <Label circular color="red">Unavailable</Label>:
+           this means you are absolutely unable to come (this will be represent by the blank spaces on your calendar).
+         </p>
+      </span>
+    );
+
     return (
       <div>
         <NavBar />
         <div className="body">
           <Card fluid raised>
-            <Card.Content textAlign="center">
-              <h5>Drag and drop to add an availability event. Click an availability event to edit.</h5>
+            <Card.Content as="h5" textAlign="center">
+                Drag and drop to add an availability event. Click an availability event to edit.
+                <Popup
+                  flowing
+                  header="Availability Info"
+                  position="bottom center"
+                  style={{ textAlign: 'center' }}
+                  trigger={<Icon name="info circle" />}
+                  content={popupContent}
+                />
             </Card.Content>
             <Card.Content>
               <div className="calendar">
